packages/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
email-validator>=2.1.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
